AbstractShape unique vertices handling + rendering improvements
Two new abstract methods added to abstractShape to support list of
unique vertices (as opposed to just some list of vertices, with
potential duplicates) to be generated. These methods are now used
for OBJ generation and rendering.

This change also enable the use of smooth normales in rendering,
which is also included in this commit.

Additionally, ambient light is now completely disabled to simulate
airless environment properly.

	modified:   abstractShape.py