Split arendConesAsteroidGenerator executable
Now it is a module instead of an executable, and is responsible for
sampling asteroid shapes only. Another module, spatialState,
samples over spatial state random variables and iterates over
their direct products with the constant ranges (phases and
distances). The new executable with the same functionality (bar the
slightly modified structure of the dataset produced) now resides at
generateArendAsteroids.py.

	modified:   arendConesAsteroidGenerator.py
	new file:   generateArendAsteroids.py
	new file:   spatialState.py